First pass at support for supporting updating facts for subscription-manager.
It caches the last found facts in /var/lib/rhsm/facts/facts.json. It trys to update facts before
a registration, or some of the list calls that go through the rules set. I'm sure there are cases where
facts should be getting updated that I've missed.

In theory, rhsmd should be triggereing facts updates as well, if they are needed.

- Adrian