Add compliance notification syslogging/desktop icon
rhsm-compliance-icon is started on desktop login. It will check on
startup and once daily via a dbus service to see if the system has any
expired entitlements, or installed products without entitlements. If so,
the user will see a notification bubble and be prompted for action.

rhsm-complianced is the backend dbus service which checks for
compliance. After a single dbus call it shuts itself down. dbus
activation is used to launch the service on any call. The daemon
does double-duty; when run with -s it will ignore dbus requests, and
just log the results of a compliance check to the syslog. This is run
via cron.nightly.