From: Adrian Likins <alikins@redhat.com> Date: Mon, 12 Jul 2010 15:23:59 -0400 Subject: [PATCH 7/7] Don't try to use any existing consumer certs for registration
A potential fix for bz#587753
Instead of using the consumber cert for the system, which can be
stale in the case of switching candlepin servers, don't use a cert
at all. Create a Restlib for the cert based auth and the basic auth
connections. Make the calls that use basic auth use it (registerConsumer
and registerConsumber getConsumerById )