Change rhms screen priority to match those of rhn.
Entitlement choose screen before, and the other after. Note, we
are probably going to need to get RHN screens to change priority,
since the eula screen and the rhn screen are both "2". Currently,
we have to skip over the EULA screen.

Make chooser screen jump to rhn or entitlement screens.