Refactor to use Python ConfigParser.
Was previously using a copied Config class from rhnclient, written years
ago at a time when Python didn't have a solid config parsing module.
This is no longer the case, and we were facing a situation where we'd
need to copy it into the UEP library RPM and sort out the deps, which
definitely does not feel like the right place. Switching to use stock
Python module.

Will require updated rhsm.conf's, see the default for an example.