Cleanup authentication logic.
Previously this was quite convoluted. UEPConnections were created with
no credentials and were thus unusable in their initial state. Certs
would have to be later added by making a separate method call on the
object. The connection also carried a separate object for basic
authentication, even though it was re-created every time it was used.

At any given time it was very hard to know what your connection object
was capable of doing and as whom.

Now, create a UEPConnection with desired credentials, and use it as
desired. These are not expensive to create, and will not longer be
morphing or in an unknown state. RHSM still carries one configured to
use the certificates on disk.