fix for bz#632612  (When re-registering, previously subscribed-to subscriptions are checked by default)
Also, fixes a few other odd behaviours that were related. We were only checking the
selected items against the "installed" itemds treelist widget, so we were throwing
gtk warnings in some cases, and not properly unselecting the row (cause of

But we were also not checking all the lists for selected items, which
was causing some odd behaviours (like not allowing you to subscribe to
a product that is installed and a product that is compatible but not
installed at the same time).

Also refactored the col_*_selected methods, and misc removal of
debug spew.