Fix broken directory path joining.
When switching to config based directories, we re-added the leading /
for paths. This broke the joining as Path.ROOT will no longer be
appended, which surfaces in Anaconda installs where it needs to be
modified.

Switched to always concatenate Path.ROOT, and if an absolute path was
given to Path.abs(), we remove the leading /.