Adjust firstboot screen priorities.
This will prevent the RHSM subscription management screen from
displaying if the user selects RHN, then declines RHN on the second
confirmation screen.

Some changes are still required here but likely in RHN code, as you
should not select RHN, then have the option to register with RHN or not.