627962: Fix issue with cpu.cpu_mhz causing facts to always update
On laptops and other machines that enable cpu speed scaling, the
fact for cpu.cpu_mhz was changing all the time, and causing our
fact's cache to always be invalid (or worse, sometimes be invalid)
and causing unneeded updating of the facts. So, add a graylist
for options to ignore changes in, and check against that before
claiming the facts have changed.

Note that "facts --update" will always send up the latest facts
regardless of the state of the cache. So if you "facts --update"
when running at 600mhz, that is what will get sent up. Later
automated checks for fact updates (say, from the daemon, or those
that happen as part of subscription related updates) will ignore
any futher changes in cpu.cpu_mhz.