Remove use of 'consumer' global in GUI.
Dangerous as it looks local, in-consistently referenced as a global
(seems to work but not sure), and causes some problems in testing as
it tries to load as soon as anything imports the module, thus very hard
to mock out.

Switching to use the Consumer object we introduced for the new UI
classes.