Remove unecessary call to server after GUI bind.
This code assumed that after a bind, we can just query the pool for
updated state, reflect this in the global add subscription screen, and
not do a full query to the server for available pools.

This is not correct however as (usually) you cannot bind more than once
to the same pool, so it should be removed entirely. Only the server will
know if this is the case or not.

a459b074a566750ac73b451921cc739b9e09a1df fixes the bug related to this,
so now this query is not even being used. (as we refresh entirely from
the server the next time the screen is opened)

Conflicts:

	src/gui/managergui.py