Fix compliance assistant dates.
Firstly, we need to be asking the server for subs valid 00:00 on the
date selected no matter what.

If unentitled products are found, the first non-compliant date needs to
show us *today*'s date.

If everything is entitled, we need the first non-compliant date to be
the first expiry end date, plus one day. (otherwise we're querying for
subs active on a date when our old sub is still active)