661517: make network config dialog respect disabling of proxy settings
Remove config settings if they are disabled with the check buttons,
but keep them around in the text fields on screen so we don't
remove them immediately on clicking the toggles.

Also, on close of proxy config dialog, update the managergui.Backend.uep
to use the new settings, so we can change proxy settings after startup.