661517: make sure changes to proxy settings are respected
This is second part of this bug fix. We were updating the config
file and object, but we had various UEP objects floating around,
and only one of them was getting updated on config changes. So
a change in the config would change the main connection, but not
say, the certlib UEP object.

So this moves everything to sharing one UEP object for the gui.
All the gui bits use the Backend.uep now.

certlib was also changed to not use it's own derived UEP class.
Methods in it were moved to the Action classes that need them.
Instead of creating a new UEP when they need it, the backend UEP
is passed down to the actions and shared.

certmgr was also previously not using proxy settings, so it now
creates one with proper proxy settings.