669753: use install -p to preserve py file timestamps
Byte compiling python files includes the timestamp of the original file
as a value in the output. As such, two files with the same content but
different timestamps will result in .pyc files with different checksums.
Install our .py files using install -p to preserve the timestamp, so
when rpm byte compiles the files, it will create the same output across
all arches, or repeated builds.