668572: hide installed subscriptions from search results
A pool for which you already have an entitlement will be returned in the
list of all pools available to your owner, but not in the list of pools
that are applicable to your system (since you already have an
entitlement, you can't be granted another one). If, however, the pool
has the multi-entitle attribute, it will always show up as being
applicable to your system (because you can get multiple entitlements for
it).

Because it does not make sense to the user to have a pool match your
hardware, then suddenly not match your hardware once you have subscribed
to the pool (we describe applicable subs as matching hardware in the
ui), hide already subscribed to pools, which are returned as being not
applicable to the system, from the search results.