670597: reload consumer in mainwindow after registration.
In firstboot, the registration and mainwindow screens are created
simultaneously at startup. Both contain a Consumer object. The
registration screen will reload it with the identity certificate
after registration, but the mainwindow instance is not updated.
Firstboot will call initializeUI method on the screen transition
which in turn calls the refresh() method on MainWindow which is
a good location for reloading its instance of Consumer.