668572: Turn search filters into real filters
Previously our filters were really toggles. Your results would either
contain only the pools applicable to your system, or they would contain only
the pools not applicable to your system.

Change the filters to be real filters, so having one of the check boxes
selected will remove results from the list, but unchecking it will not.
Start with all filters turned on, meaning we only display the most
appropriate pools for the system.

To make all of the checkboxes behave the same, some of the label wording
had to change (becoming more verbose). To fit them in the UI, they were
moved from a horizontal to a vertical orientation.