670655 can't dismiss error dialog when subscribing to personal subscriptions
Something is scrambling a reference, but I cannot figure out where it is. This
patch keeps the reference sane, but it technically should be an operation with
no side effects.

For example, if you print the "self" ref in both the callback and constructor
*WITHOUT* the hook, you get this (incorrect):

<ErrorDialog object at 0x7fbe582cb5f0 (gui+messageWindow+MessageWindow at 0x32d0360)>
<ErrorDialog object at 0x7f7b242cc550 (uninitialized at 0x0)>

However, if you add the hook, you get this (correct):

<ErrorDialog object at 0x7f87d82cc640 (gui+messageWindow+MessageWindow at 0x2945700)>
<ErrorDialog object at 0x7f87d82cc640 (gui+messageWindow+MessageWindow at 0x2945700)>

I stepped through this with a debugger, but I couldn't find where the object
switcharoo was happening.