676377: rhsm-compliance-icon's status can be a day out of sync
We use D-Bus to call rhsm_compliance_d on certain events. This fires
a signal that rhsm_compliance_icon listens for. The signal already contains the
new status, so there is no need for a second remote call.

-------------
|  CLI/GUI  |--.
-------------   \   ---------------------
                 `> | rhsm_compliance_d |
    calls           ---------------------
    check_compliance   ^
                      /  listens for
                     .   signal
                    /
-------------------------
| rhsm_compliance_icon  |
-------------------------

Note that calls can never go FROM rhsm_compliance_d, because it is
on the system bus and not session bus.