682311: rhsm-compliance icon pegs the cpu at 100%
The rhsm_complianced script is normally invoked via dbus, in a fashion similar
to an xinetd script. It runs once, does the needful, and then quits.

To ensure that rhsm_complianced quits, there was a "watchdog" added via
idle_add on the main glib loop. Previously, it was invoked at startup, and
polled until the script's dbus invocation was complete, then forced termination
of the main loop.

It appears that during firstboot, there is a way for rhsm_complianced to be
executed via dbus in a way that made the watchdog activate, but did not
actually perform a call to any remote method. This made glib consume 100% cpu,
since the watchdog would poll forever.

This fix simply moves the watchdog invocation later into the script, to shorten
the polling window to more closely match the actual execution of the remote
method.

The firstboot code is still invoking the remote interface improperly, but
that's a different bug with a different priority. It seems to have no
functional effect, aside from a lingering process which gets reaped after a
successful compliance check call.