681925: subscription-manager masks SIGPIPE when running virt-what, resulting in errors in shell commands
A user reported via the beta list that he was seeing "Broken pipe" errors when
running virt-what via subscription-manager. It turns out that Python masks
SIGPIPE for itself, but it fails to reset the mask when forking. See
http://bugs.python.org/issue1652 for more detail.

This patch (which was graciously provided by rjones) manually alters the
sigmask to not ignore SIGPIPE before forking, and then resets the mask back
after the execve().