688192: don't look for dmi info on machines without dmi
dmi/smbios information isn't available on all machines that run linux.
Notably, ppc and s390 do not support it. Detect if we're running on a
ppc, s390, or related arch, and if so, do not try to load dmidecode and
read dmi info for hardware facts.

Since we're not trying to read dmi for these machines, skip
python-dmidecode as a dep for them, too.