694837: Fix handling of unsubscribing imported certs while unregistered
The core of the bugfix is in mysubstab._on_unsubscribe_prompt_response.
We don't try to unbindBySerial if we are not registered (via new
managerGui.Backend.is_registered()). We also try/except around
the update to catch the Disconnected exception if well, we are
disconnected. We also change the store serial type to long, so
the cert lookup works correctly.

we also remove some unused code from certlib/certmgr (the *add) code
and refactor certlib.delete to expect an array of longs, not a
var arg.