694851: only open one dbus proxy connection
On EL 5 we were getting a failed assertion in the dbus code. One of our proxies
got into a bad state where it was created, but not marked as being associated
with a backing dbus connection. When dbus signaled our app that the owner of
the com.redhat.SubscriptionManager endpoint changed (probably due to dbus
activation), this unassociated proxy was noticed by dbus, and the assertion
failed, crashing the app.

This would happen predictably for the first few times you log in, then stop
(not sure why).

Using a single proxy connection seems to fix it. My theory is the signal
listener was the proxy that would die, its death being caused by dbus
activation changing the signal controller. With one proxy, both the signal
listener and the dbus caller use the same proxy, so it always knows what's
going on.