696020: on x86, serial numbers do not like to be long's
Trying to pass a way longer than 4byte 'long' into gtk tree view label
was throwing errors, so just use it as a string and cast.

Also, the method we were using to detect if we were registered
wasn't working if no admin api calls were made, so just check
for existence of consumer id.