696791: Change the way we detect cpu sockets on s390x
On RHEL5.7, the topology info we were depending on in the cpu
socket calculation just doesn't exist, so it was failing.

On RHEL6.x, the topology info exists, but only for one cpu
entry, so it was failing.

Also, "sockets" doesn't really exist on s390x, so it was failing.

On RHEL6, we were using lscpu output sometimes, which is slightly
different on s390x, so it was failing.

So now, we do s390x specific cpu info collection, and basically
return the number of available cpu's as "cpu.cpu_socket(s)" and
the number of "online" cpu's as "cpu.cpu(s)"