700547: ensure the notification attaches to rhsm-icon
For el5, we have to wait to show the notification bubble until the
systray icon is both created, and visible on the screen. Otherwise the
notification daemon isn't able to properly position it. To do this, we
create the icon at application startup, and just toggle its visibility,
and add a callback that waits for the icon to send a size-changed
signal, meaning it's now visible on the screen, before displaying the
notification.