700958: fire check_status dbus calls in a non-blocking manner
From the subscription-manager cli and gui, we don't really care about
the response from the check_status dbus call. From the cli, the dbus
call was blocking, causing two problems:
 - the cli command wouldn't exit until the daemon finished processing
   the call.
 - if the daemon took too long to respond, we'd get a timeout exception,
   even though the daemon could still complete sucessfully.

Using a non-blocking call from the cli would be most ideal, but it's not
easily doable from both el 5 and el 6. For now, keep the call as
blocking, but if we catch an exception from a timeout, just swallow it
so as to not alarm the user.

NOTE: to test, add a time.sleep(some_value_greater_than_30) to
check_status in rhsm_d.py, and run the daemon from source with -d -k