710458: Do not display activation messages as errors.
While still throwing a RestlibException, activation messages are
actually coming back with HTTP status 202 (accepted) and should not be
rendered as errors.

Instead we'll watch for this in the generic exception handling methods
and ignore the error message when we see one, instead showing an info
dialog with just the translated message from the server.