700547: delay running the first icon status check
On rhel 5 and in VMs, we're seeing the notification bubble from the icon
not rendering in the correct location when launched from gnome-session
(on login). Default to running the first check 4 mins after launching,
which should give gnome-panel enough time to load up, so the bubble can
be positioned correctly.

Add a new cli option --check-immediately (-i) to fire the check right
away, for testing.