706265: Try to detect when removed pkgs make a repo not "active"
This is a workaround for lack of metadata containing the info
about what repo a package was installed from.

We look at our current metadata, and try to figure out what repo
that package would come from now, and consider those repo's equal
to the repo that it was installed from.