711133: Handle updates from old style key.pem certs
If we don't have a $SERIAL-key.pem associated with each
cert file, consider that cert invalid, and try to update it.

For the upgrade case, this means we write it with the new style
Writer(), and get the new style $SERIAL-key.pem file format. So
we can use this to migrate from old style certs to new style