Fix backward compatability with old use of getPoolsList.
New changes failed to account for existing code calling the method,
which now breaks, but we could do and still maintain backward
compatability cleanly.

Fix method to allow calling with just a consumer UUID, use the new URL
structure if an owner is optionally specified. (no longer required) Also
can now list pools for just an owner if so desired.