Select an environment during registration if supported.
Checks on the server to see firstly if environments are supported. If so
we check if there is just one available and use it automatically. If
multiple are available we prompt the user.

Uses similar mechanism to the owner selection, but follows it as you
must first select an owner before we can even list the environments.