Remove a pointless Action hierarchy.
There are about three Action hierarchies floating around the code base,
all of which are disjoint and do different things but look extremely
similar. This one appears completely unnecessary as we're already
abstracting behind the 'lib' object (factlib in this case), and there's
no need for a factlib object doing nothing.