706889: Fix -1 printed from CLI when exiting with Python 2.4.
Issue reportedly only surfaces in RHEL 5. Originally this fixed unbind
with a bad serial from printing the -1 after the error message, but the
fix likely corrects a number of other CLI sys.exit's as well.

Print was originating in the handle_exception method, change it to check
if the type is a SystemExit, which implies we already told the program
to exit and thus, don't want to print any messages or do any handling.