Upload info about the installed products on the system.
Refactor some cached JSON with server update logic into a base class,
and added a new implementation for sending installed products data to
the server and caching what was last sent for comparison. Ideally we'll
use this for facts someday too.

Rename pkgprofile module to cache. Don't want to keep adding new modules
for single classes, so switching to a more generic cache module for each
of these pieces of consumer data we push to the server when something
has changed.

Pass up installed products during registration, and when rhsmcertd runs.
(if anything has changed)