implement red/yellow/green for the cli
managerlib when through some refactoring to support r/y/g. Instead of
going directly to the entitlement and products directories, it was changed
to utilize cert_sorter. As a result of this change managerlib now looks
at partially installed products, expired, unentitled, and installed.

Enhanced cert_sorter to track not installed products, these are products
where we have an entitlement certificate but no product certificate.
cert_sorter was also changed to support multiple entitlements per
product_id, resulting in the valid_products, uninstalled_products,
to have a list of entitlements vs just a single one.

The stacking test case was modified to look at the valid_products not
the unentitled_products to support the changes to cert_sorter.