Client-side changes for healing
We look for an "autoheal" directive in rhsm.conf. If it exists, then when
rhsmcertd is invoked, it will perform an autosubscribe call using "bind"
instead of "bindByProduct".

For now, the SubscribeCommand uses this new style of bind, and RegisterCommand
still uses the old one.  This is a compromise since the new bind will be in CP
shortly, but isn't currently available. Registration with --auto (the most
common use case) will still work, but subscribe --auto will fail until the
server-side changes are implemented. Once the server-side work is complete, we
can alter RegisterCommand as well.