Use server side consumer autoheal flag.
Rather than the client side config option we originally implemented.
This is due to a new requirement that the setting be editable in a
management webUI.

Moving healing to a "lib" action run by certmgr. Allows us to ensure
that before we heal, we sent up facts and installed products, and after
we heal we refresh certificates.

Switched from re-using the CLI subscribe command to an explicit call,
this avoid doing things like fact/certificate updates multiple times per
run.

Removed certlib passing to autosubscribe method, most callers do this
after the autosub call. (fixed the one case that wouldn't)