Add a "refresh" method to cert_sorter
There was a bug in myinstalledtab where you could be green, go to yellow, and
go back to green without the product list icons changing back to green as well.
The underlying issue was that cert_sorter was not being refreshed. I broke out
some of the cert_sorter __init__ into a refresh() method, which is callable
without having to pass in all the c'tor params again.