734880: Handle bundled certs in the installed produict status.
I am not sure if this accounts for yellow. However, the listing code was asuming that all the products in a cert were either installed or not. This is not hte case for bundles. The bundle can have some installed, and some not. If you have 2 certs which give you a product, this will still list hte product twice. Once per cert.