736166: move certs from subscription-manager to python-rhsm
The RHEVH folks are using python-rhsm sans subscription-manager. In order to
communicate with hosted CP, they need the certs that have historically been
bundled with sub-mgr. I moved them out into python-rhsm, and removed the
fakamai cert that I believe is no longer in use.

There is another change in the subscription-manager repo to remove the certs
from the subscription-manager package.