Changes to rhsmcertd to support healing frequency (part I).
A new rhsm.conf var for "healFrequency" was defined. This is read from the init.d script,
and passed in as a second arg to rhsmcertd.

Note that healFrequency is mandatory, and that it doesn't guarantee that a heal
will be performed. The following criteria still need to apply for an autoheal:

* the system must not be in "green" status
* the server-side flag for autoheal must be set

Part II will move from the current one-loop-per-job impl to using a shared
event loop (most likely glib's). I need to verify that it's OK to add this dep,
so I'm using the simple implementation for now while that gets sorted out.