737841: Handle dates beyond 2038 on 32-bit systems.
If a subscription expires beyond Jan 19 2038, 32-bit systems would error
out and be unable to list subscriptions. (and probably do a lot of other
things)

If we encounter such an overflow error, replace the date with Jan 1 2038
(which will not overflow). Such a system will clearly be in big trouble
on that date, but this allows them to continue listing subscriptions
today which might have start/end dates beyond 2038.