738549: Allow subscription-manager to run without dbus
The subscription-manager cli fires off a dbus message whenever it
modifies your subscription status. If dbus is not running, this will
fail, and display a nasty message to the user. Since we allow the
message itself to be fired without caring what happens to it, allow the
same for acquiring a dbus connection.