740046: Ensure common behavior on dates between CLI and GUI
The dates for begin and end for the installed option will be based on the
same information for both CLI and GUI. No entitlements for an installed
product results in a blank for both begin and end dates.

Expired, future, partial, and full subscriptions will go with a compiling
of dates from all availble entitlements for a given product.