Add logging, tests, and comments for broken yellow detection.
This commit adds a number of tests (commented out) which expose broken
behavior in the yellow status detection. Today this only works in a very
narrow situation which is unlikely to match up with real world data. As
such yellow detection is best considered broken, but we did not have
time to fix for next release. See comments in cert_sorter.py for
explanation of the problem.

Added some logging to cert sorter to help understand what's going on.

Heavy refactoring of the old tests, many were not properly testing what
they were intended too, and those that were were checking only the
narrow case. (which works) Several were removed/replaced.