692242: rhsm_icon disappears and will not return
If rhsm_icon is running and goes into the partial/invalid state, they will see
the icon. If they get into the valid state, they will then have the icon
disappear. However, if they then go BACK to the partial and/or invalid state,
the icon will NOT reappear. This is because previously, the GtkStatusIcon was
being destroyed via a reference to g_object_unref() in destroy_icon(). The
GtkStatusIcon was only created during rhsm_icon startup, and was not re-created
inside of alter_icon(). Thus, calls to context->icon to redisplay the icon were
bombing out.

Rename destroy_icon() to hide_icon(), and do the actual GtkStatusIcon
destroying with the other cleanup activities at the end of main().