740675: do a condrestart on rhsmcertd when we upgrade subscription-manager
Recently, we added a healing option to rhsmcertd. However, if a user does an
RPM upgrade instead of a fresh install, they will get the new rhsmcertd binary
laid down without a restart. The user then has to manually restart rhsmcertd to
pick up the change.

Implement a condrestart in the init.d script that only restarts rhsmcertd if it
is already running. Then, have the spec file perform a condrestart in the %post
only in the case of upgrades.