740773: Do not delete certs if we have repo metadata errors
products cert's installed by anaconda were getting deleted
on first product install by yum (via product-id plugin)
if we have repo metadata issues (in this case, a checksum
mismatch for the product file).

product-id plugin was not able to create the proper
cert->repo mapping in productid.js because of the
chksum errors. So the productid.js had the cert
associated with a repo name that only exists in anaconda.
This repo appeared "inactive", and then would be
deleted.

This fix prevents any cert deletions from happening
if the metadata presents any errors.