742013: sub-mgr translations not showing up in firstboot
Firstboot was not showing translated strings inside the rhsm_login.py and
rhsm_subscriptions.py, but only for strings that were set via glade's
"translatable" flag, and not via the usual _().

The solution was to set glade.gtk.textdomain() explicitly in registergui.py.
Setting this via configure_i18n() did not work for some unknown reason,
possibly due to the way we interact with the RHN firstboot module.

Thanks to alikins, who was the one who discovered this.