742416: Remove the close button from our progress dialog
Our async code, which runs behind the progress dialog, is not
cancelable in any way. As such, when the progress dialog is closed via
the window 'x' button, we throw an exception and keep running the async
job to completion. the exception is usually swallowed, but in firstboot,
it's visible.

Since we can't really cancel the async work, just get rid of the
progress dialog's close button.