737145: prevent whitespace from building up in redhat.repo
When a repo config changed, we'd remove its section, then readd it, to
get rid of any deleted config values. Removing a section means you end
up with an area in your repo file with two consecutive newlines. as time
passed, these newlines would build up, and gather at the top of the repo
file. Fix this by just removing the values in the section, rather than
the whole section itself.

Additionally, remove the Reader class, which was supposed to take care
of double newlines, but didn't, and replace it with iniparse's tidy
function.