743704: Fix healing date issues.
Healing code was written to re-use find_first_invalid_date, but this
originated in the GUI where we actually need to find the first date
where we're *completely* invalid at midnight GMT.

This causes the healing code to think we're not yet compliant if it
happens to run in-between.

Instead switch healing to simply do cert sorting for todays date and
tomorrows, which gives us reliable results and technically is quicker,
as the cert sorter could be used heavily in find_first_invalid.

Adding comments for the odd behavior in find_first_invalid_date. This
probably should be cleaned up some day but it's a delicate balance to
get the behavior we need.

Conflicts:

	src/subscription_manager/certlib.py
	test/test_cert_sorter.py