Refresh certificates after healing.
Normal configuration has healing followed by a cert refresh, but if the
user were to have altered the normal intervals healing could run by
itself. In this scenario the certificates would not actually get pulled
down until the next cert refresh, potentially defeating the purpose of
the healing in the first place.

This has the unfortunate side effect of an extra cert refresh request
when the daemon runs. This will happen frequently with default config,
but bear in mind it is just a cert serial comparison before any actual
cert content will be pulled down.