Remove post-healing logging of product status.
This has always been broken as certs weren't refreshed after healing,
and the output was also broken as the CLI was recently broken. (the
method used started returning tuples with different indexes)

The logging during cert refresh should be sufficient for what is
intended here, and possibly much less verbose.