Fix a test case that failed to take DST into account.
The datetime we were generating to compare the result of
ParseDate to was not created with a timezone, so DST is
not being applied, so take this into account.

Also, some mild refactor to fix some datetime.* usages
that were confusing. we were importing datetime from
modelhelpers namespace, and that datetime was a
datetime.datetime class, making the names weird.