Cleanup auto-subscribe behavior.
GUI register with auto-subscribe was still using the old style of
submitting installed product with the auto-subscribe request, rather
than the new style for healing which assumes the server already knows.

Also removes a temp hack in the CLI for maintaining compatability while
new versions of Candlepin rolled out. This should no longer be needed.