705883: Fix error dialog modal issues.
Error dialogs were not getting a parent window set, so it was possible
to click any other open part of subscription manager and have it hide
the error dialog, leaving you with an unresponsive window in the
foreground.

This commit requires all callers of handle_gui_exception to now specify
the parent window the error dialog should be associated with. This will
lock the dialog to that window and prevent anything from being placed
over top of it.

errorWindow also now accepts parent argument.

Required some of the main window tabs to now have a reference to the
main window itself, where they previously did not.