756507: do not use output from "getlocale" as input for "setlocale"
In RHEL5 and possibly scenarios in RHEL6 that we are unaware of, the output of
locale.getlocale() is not considered valid input for locale.setlocale().

Instead of using getlocale(), use an empty string which will reset the locale
to the system default. See http://bugs.python.org/issue1699853 for more detail.
Also, use 'en_GB' sans encoding when resetting to the safe date parsing locale.