740788: Getting error with quantity subscribe using subscription-assitance page.
Altered process so that all of the pool group is used when selected. When the page says 12 of 15 as available [a number that is a sum of the pools] it can use all 12 across the pools.

Also included a limiter that only allows the quantity box to equal the number available when using the up arrow. The subscription assistant also has a check in case the user types a vlue higher than the number available.

Other changes discussed to contract selection via all subscriptions screen will be completed later.