759069: catch exception when enabling invalid repositories
One of the steps of the migration script is to perform an autosubscribe, and
enable any extra repositories (namely rhel-$ARCH-$VARIANT-supplementary-5).
However, if the autosubscribe was unable to cover any of the user's products,
then enabling the extra repositories will fail since there is no available
additional content to enable.

This patch wraps the extra repository logic in a try/except block, to give a
better error message to the user in the event of an autosubscribe failure. A
cleaner implementation would be to have subscription-manager return a non-zero
return code when autosubscribe is not able to fully cover the user's
entitlements and then check for that code in the migration script, but I'm not
comfortable doing that right now since it would require changes to the
migration script, managercli.py, and the automation.