Fix a test case failure that only happens Jan-Sep
We have a test to check if our workaround for strptime failures
is still needed. For ja_JP and ko_KR, strptime works on single
digit months, so change the test to expect it to work for those
dates. Also, stop testing "now" and use fixed dates so this
behaviour is more consistent.