767620: Add manage_repos config option.
When set to 0, will disable generation of the redhat.repo file. The
redhat.repo file will be deleted on the next rhsmcertd run and should
not be re-created as long as this config value is set.

subscription-manager repos --list will print "Repositories disabled by
configuration."

If the config option is unset or 1, we resume normal behavior.

squash