Have client check sockets on non-stacked entitlements as well.
This commit mirrors the changes made to the server's compliance checking
rules.

If a consumer has non-stacked entitlements with a socket limit that is
less than the number of sockets the consumer has, this should be treated
as a partially valid product.

If any non-stacked entitlement has enough sockets, we will still
consider them to be fully valid.

If multiple non-stacked entitlements do not have enough sockets, they do
*not* get added together, the system would still be considered partially
valid.

Quantities are considered, as the subscription could still allow
multiple entitlements.