Set and use the system's service level.
When we select a service level we should be setting it on the server.

If we enter the autobind wizard already having a service level defined
on the server, we should use it instead of promoting the user to select
another one.

In this scenario, we consider it ok to proceed even if the current SLA
will not get you entitlements for every product needing them.