752756: Cache the facts, and refresh the validity facts whenever they change.
The bug was caused by the fact that some of the hardware facts are determined by calling out to
virt-who which is wrapped by signal calls. Signals can only be used in the main loop. When the
fact checking was done outside of the main loop then an exception is thrown. This exception causes
the consumer's virt type to be marked as Unknown. The Unknown status causes virt pools to not be
brought down to the client.