800121: do not attempt to call UEP when system is unregistered
During unregistration, one of the steps is to run through an update on the
system in order to clear out any data that is no longer applicable. A step in
the update list is to fetch the list of expected certs from candlepin, which
will not work since the system is no longer registered.

Instead, simply return an empty list for expected certs if UEP object is not
initialized.