Update CLI to handle server that doesn't support service levels.
Tried to do this without any additional server requests, but had to add
one in the case if subscribe --auto.

If the server does not appear to support service levels (usually by
checking the consumer's JSON for a serviceLevel attribute), we error
out. Any previous actions are not undone. (i.e. during register, you
will remain registered, just not autosubscribed)