812446: Fix deletion of productids with yum localinstall
If we encounted a scenario where no repo's were considered
"enabled" and no "active" repos, we would delete the productid.
A fresh install running a "yum localinstall" would trigger
it, since no repo's have been enabled.

Check to verify we are not in this scenario before deleting
productid's.

Also, add some logging describe which productids are being
deleted. Previously we had no longer for this.