804201: Fix sla select in firstboot after back button
If you moved past the sla select screen, then hit back to return to it,
the list of SLAs would not be drawn until you tabbed into them.

This appears to have been caused the by gtk alignment in the widget
hierarchy. Remove it, and use pack_start to order the slas starting from
the top of the screen rather than relying on the alignment to push them
to the top.